+++ android/src/main/java/com/getcapacitor/community/database/sqlite/SQLite/UtilsSQLite.java
@@ -3,12 +3,15 @@ package com.getcapacitor.community.database.sqlite.SQLite;
 import android.util.Log;
 import androidx.sqlite.db.SupportSQLiteDatabase;
 import com.getcapacitor.JSArray;
+import java.nio.ByteBuffer;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import net.sqlcipher.Cursor;
 import org.json.JSONArray;
 import org.json.JSONException;
+import org.json.JSONObject;
 
 public class UtilsSQLite {
 
@@ -90,7 +93,7 @@ public class UtilsSQLite {
         Object[] objectList = listArray.toArray();
         String[] retArray = Arrays.copyOf(objectList, objectList.length, String[].class);
 
-//        String[] retArray =  listArray.toArray(new String[listArray.size()]);
+        //        String[] retArray =  listArray.toArray(new String[listArray.size()]);
         return retArray;
     }
 
@@ -108,7 +111,7 @@ public class UtilsSQLite {
 
     private List<String> trimArray(List<String> listArray) {
         List<String> trimmedStrings = new ArrayList<String>();
-        for(String s : listArray) {
+        for (String s : listArray) {
             trimmedStrings.add(s.trim());
         }
         return trimmedStrings;
@@ -120,7 +123,17 @@ public class UtilsSQLite {
             if (jsArray.isNull(i)) {
                 list.add(null);
             } else {
-                list.add(jsArray.get(i));
+                Object obj = jsArray.get(i);
+                if (obj.getClass() == JSONObject.class) {
+                    if (((JSONObject) obj).getString("type").equals("Buffer")) {
+                        byte[] bArr = JSONArrayToByteArray(((JSONObject) obj).getJSONArray("data"));
+                        list.add(bArr);
+                    } else {
+                        throw new JSONException("Object not implemented");
+                    }
+                } else {
+                    list.add(obj);
+                }
             }
         }
         return list;
@@ -139,11 +152,40 @@ public class UtilsSQLite {
         return list;
     }
 
+    public byte[] JSONArrayToByteArray(JSONArray arr) throws JSONException {
+        byte[] bArr = new byte[arr.length()];
+        for (int i = 0; i < arr.length(); i++) {
+            bArr[i] = (byte) (((int) arr.get(i)) & 0xFF);
+        }
+        return bArr;
+    }
+
     public Boolean parse(Object mVar) {
         boolean ret = false;
         if (mVar instanceof JSONArray) {
             ret = true;
         }
         return ret;
+    }
+
+    public int ByteToInt(byte BVal) {
+        String comb;
+        int out = 0;
+        comb = BVal + "";
+        out = Integer.parseInt(comb);
+        // Get Unsigned Int
+        if (out < 0) {
+            out += 256;
+        }
+        return out;
+    }
+
+    public JSArray ByteArrayToJSArray(byte[] BArr) {
+        JSArray arr = new JSArray();
+
+        for (int i = 0; i < BArr.length; i++) {
+            arr.put(ByteToInt(BArr[i]));
+        }
+        return arr;
     }
 }