+++ ios/Plugin/Extensions/Data.swift
@@ -1,0 +1,13 @@
+//
+//  Data.swift
+//  CapacitorCommunitySqlite
+//
+//  Created by  Quéau Jean Pierre on 10/06/2023.
+//
+import Foundation
+
+extension Data {
+    var bytes: [UInt8] {
+        return [UInt8](self)
+    }
+}